21	Feature	New	Normal	Add Change Login/Key Support
There should be a workflow where the user either can change their site
key and by an email request to we locally or, that we can tell them to
change their site key by providing it to them. The goal of this is to
have 100% control of the subscribe process and to be able to back door
anybody we want.