Bug #65 Fix workflow issues for Place Finder
The place finder is a wizard, it has a workflow. The primary workflow
is to try to find a place and if you can't find it you can add it. The
flow of next and back is not fully implemented. Work needs to be done
to provide the full wizard workflow.